Update operation, clear, and sign change buttons
The operation buttons were updated so they properly perform the
functions assigned to them with a switch statement when the equal
button is pressed.

The clear button was updated to change firstNum, secondNum, and,
operation to null. The sign change button properly changes either the
firstNum or secondNum to the opposite of positive or negative.